+++ engine/apps/slack/scenarios/distribute_alerts.py
@@ -23,7 +23,6 @@ from apps.slack.client import (
     SlackClientWithErrorHandling,
 )
 from apps.slack.constants import CACHE_UPDATE_INCIDENT_SLACK_MESSAGE_LIFETIME, SLACK_RATE_LIMIT_DELAY
-from apps.slack.models import SlackMessage
 from apps.slack.scenarios import scenario_step
 from apps.slack.scenarios.slack_renderer import AlertGroupLogSlackRenderer
 from apps.slack.slack_formatter import SlackFormatter
@@ -136,16 +135,12 @@ class AlertShootingStep(scenario_step.ScenarioStep):
         try:
             result = self._slack_client.chat_postMessage(channel=channel_id, attachments=attachments, blocks=blocks)
 
-            slack_message = SlackMessage.objects.create(
+            alert_group.slack_messages.create(
                 slack_id=result["ts"],
                 organization=alert_group.channel.organization,
                 _slack_team_identity=slack_team_identity,
                 channel_id=channel_id,
-                alert_group=alert_group,
             )
-
-            alert_group.slack_message = slack_message
-            alert_group.save(update_fields=["slack_message"])
 
             # If alert was made out of a message:
             if alert_group.channel.integration == AlertReceiveChannel.INTEGRATION_SLACK_CHANNEL:
@@ -156,13 +151,12 @@ class AlertShootingStep(scenario_step.ScenarioStep):
                     text=":rocket: <{}|Incident registered!>".format(alert_group.slack_message.permalink),
                     team=slack_team_identity,
                 )
-                SlackMessage(
+                alert_group.slack_messages.create(
                     slack_id=result["ts"],
                     organization=alert_group.channel.organization,
                     _slack_team_identity=self.slack_team_identity,
                     channel_id=channel,
-                    alert_group=alert_group,
-                ).save()
+                )
 
             alert.delivered = True
         except SlackAPITokenException:
@@ -823,7 +817,6 @@ class AcknowledgeConfirmationStep(AcknowledgeGroupStep):
             )
 
     def process_signal(self, log_record: AlertGroupLogRecord) -> None:
-        from apps.slack.models import SlackMessage
         from apps.user_management.models import Organization
 
         alert_group = log_record.alert_group
@@ -886,13 +879,12 @@ class AcknowledgeConfirmationStep(AcknowledgeGroupStep):
                 else:
                     raise e
             else:
-                SlackMessage(
+                alert_group.slack_messages.create(
                     slack_id=response["ts"],
                     organization=alert_group.channel.organization,
                     _slack_team_identity=self.slack_team_identity,
                     channel_id=channel_id,
-                    alert_group=alert_group,
-                ).save()
+                )
 
                 alert_group.slack_message.ack_reminder_message_ts = response["ts"]
                 alert_group.slack_message.save(update_fields=["ack_reminder_message_ts"])
@@ -986,10 +978,7 @@ class UpdateLogReportMessageStep(scenario_step.ScenarioStep):
         self.update_log_message(alert_group)
 
     def post_log_message(self, alert_group: AlertGroup) -> None:
-        from apps.slack.models import SlackMessage
-
-        slack_message = alert_group.get_slack_message()
-
+        slack_message = alert_group.slack_message
         if slack_message is None:
             logger.info(f"Cannot post log message for alert_group {alert_group.pk} because SlackMessage doesn't exist")
             return None
@@ -1024,15 +1013,13 @@ class UpdateLogReportMessageStep(scenario_step.ScenarioStep):
                     raise e
             else:
                 logger.debug(f"Create new slack_log_message for alert_group {alert_group.pk}")
-                slack_log_message = SlackMessage(
+                slack_log_message = alert_group.slack_messages.create(
                     slack_id=result["ts"],
                     organization=self.organization,
                     _slack_team_identity=self.slack_team_identity,
                     channel_id=slack_message.channel_id,
                     last_updated=timezone.now(),
-                    alert_group=alert_group,
                 )
-                slack_log_message.save()
 
                 alert_group.slack_log_message = slack_log_message
                 alert_group.save(update_fields=["slack_log_message"])
@@ -1044,8 +1031,7 @@ class UpdateLogReportMessageStep(scenario_step.ScenarioStep):
             self.update_log_message(alert_group)
 
     def update_log_message(self, alert_group: AlertGroup) -> None:
-        slack_message = alert_group.get_slack_message()
-
+        slack_message = alert_group.slack_message
         if slack_message is None:
             logger.info(
                 f"Cannot update log message for alert_group {alert_group.pk} because SlackMessage doesn't exist"